// packages/types/src/enums/issueType.ts
/** All issue categories a reporter can select when contacting a vehicle owner. */
export enum IssueType {
  BLOCKING_VEHICLE = 'BLOCKING_VEHICLE',
  WRONG_PARKING = 'WRONG_PARKING',
  LIGHTS_ON = 'LIGHTS_ON',
  DOOR_OPEN = 'DOOR_OPEN',
  FLAT_TYRE = 'FLAT_TYRE',
  FLUID_LEAKING = 'FLUID_LEAKING',
  VEHICLE_DAMAGE = 'VEHICLE_DAMAGE',
  EMERGENCY = 'EMERGENCY',
}